+++ Singular/weight.h
@@ -6,7 +6,7 @@
 /*
 * ABSTRACT
 */
-/* $Id: weight.h,v 1.4 2000-10-23 12:02:23 obachman Exp $ */
+/* $Id: weight.h,v 1.5 2000-10-30 13:40:29 obachman Exp $ */
 
 #include "structs.h"
 
@@ -17,8 +17,8 @@ extern pLDegProc pLDegOld;
 void kEcartWeights(polyset s, int sl, short *eweight);
 BOOLEAN kWeight(leftv res,leftv id);
 BOOLEAN kQHWeight(leftv res,leftv v);
-int maxdegreeWecart(poly p,int *l, ring r = currRing);
-int totaldegreeWecart(poly p, ring r = currRing);
+long maxdegreeWecart(poly p,int *l, ring r = currRing);
+long totaldegreeWecart(poly p, ring r = currRing);
 
 short * iv2array(intvec * iv);
